+++ src/main/java/com/CPR/redHome/mapper/admin/product/ProductAdminMapper.java
@@ -17,13 +17,13 @@ public interface ProductAdminMapper {
     List<ProductDto> selectProducts(@RequestParam int firstRecordIndex, @RequestParam Criteria criteria);
 
     // productId로 상품 조회
-    ProductDto selectProductByProductId(int productId);
+    ProductDto selectProductByProductId(Long productId);
 
     // 상품 update
     void updateProduct(ProductDto productDto);
 
     // 상품 delete
-    void deleteProduct(int productId);
+    void deleteProduct(Long productId);
 
     // 가격대 별 상품 수
     LinkedHashMap<String, Integer>  selectProductByPrice();
